'''Attach a RandomGenerator to each ExecutionContext'''
 * ExecutionContext: contains now a RandomGenerator
 * Bandits: remove RandomGenerators from Bandit states and policies
 * Every body now uses the RandomGenerator that is attached to the current ExecutionContext

+++ trunk/projects/Proteins/Predictor/ProteinPerception.h
@@ -537,7 +537,7 @@ public:
     const size_t n = cysteinIndices.size();
     VectorPtr res = vector(positiveIntegerType, n);
     for (size_t i = 0; i < n; ++i)
-      res->setElement(i, Variable(abs(cysteinIndices[i] - position), positiveIntegerType));
+      res->setElement(i, Variable(abs((int)cysteinIndices[i] - (int)position), positiveIntegerType));
 
     return res;
   }
